Soft/requirements.txt
# ─── Core Web Framework ───────────────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
Werkzeug==3.0.3

# ─── Database ─────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9      # PostgreSQL driver (swap for SQLite in dev)
# For SQLite dev, no extra driver needed

# ─── Face Recognition: MTCNN + FaceNet ────────────────────────────────────────
facenet-pytorch==2.6.0       # Includes MTCNN + InceptionResnetV1 (FaceNet)
torch==2.3.0
torchvision==0.18.0

# ─── Image / Video Processing ─────────────────────────────────────────────────
opencv-python==4.10.0.82
Pillow==10.3.0
numpy==1.26.4

# ─── Security & Auth ──────────────────────────────────────────────────────────
PyJWT==2.8.0
bcrypt==4.1.3
cryptography==42.0.8

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
marshmallow==3.21.3
pydantic==2.7.1
python-multipart==0.0.9
requests==2.32.3
scipy==1.13.0                # cosine similarity for embedding comparison
